Improve ability to install outside of PG
I've updated comments and placed some hardwired assumptions
that were fine for us into new variables.  With the new changes
we should be able to move the directory name without errors.
We'll be changing the name of the repository, so that will be
required.

I also made revisions to bring us closer to being able to have
a UI to edit an installed module's name, etc as these first
modules may be more likely to act as a template until we upload
a template repository.